@@ -1,13 +1,15 @@
 {
   "name": "uiv",
   "version": "1.4.1",
-  "description": "Bootstrap 3 components implemented by Vue 2.",
+  "description": "Bootstrap 3 components implemented by Vue.",
   "author": "wxsm",
   "main": "dist/uiv.common.js",
   "module": "dist/uiv.esm.js",
   "unpkg": "dist/uiv.min.js",
   "jsdelivr": "dist/uiv.min.js",
-  "files": ["dist"],
+  "files": [
+    "dist"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/uiv-lib/uiv.git"
@@ -20,8 +22,8 @@
   "scripts": {
     "start": "npm run dev",
     "dev": "npm run docs:dev",
-    "docs:dev": "vuepress dev docs",
-    "docs:build": "vuepress build docs",
+    "docs:dev": "vitepress dev docs",
+    "docs:build": "vitepress build docs",
     "dist": "npm run dist:esm && npm run dist:esm-single && npm run dist:cjs && npm run dist:umd",
     "dist:umd": "rollup -c build/rollup.umd.js",
     "dist:cjs": "rollup -c build/rollup.cjs.js",
@@ -31,24 +33,24 @@
     "build": "node build/build-docs.js && node build/build-dist-config.js",
     "test": "jest --coverage",
     "test:watch": "pnpm test -- --watch",
-    "lint": "eslint --ext .js,.vue src build docs/.vuepress",
+    "lint": "eslint --ext .js,.vue src build docs/.vitepress",
     "lint:fix": "pnpm lint -- --fix"
   },
-  "dependencies": {
-    "portal-vue": "^2.1.7",
-    "vue-functional-data-merge": "^3.0.0"
-  },
   "devDependencies": {
+    "@babel/preset-env": "^7.16.0",
     "@rollup/plugin-alias": "3.1.8",
     "@rollup/plugin-buble": "0.21.3",
     "@rollup/plugin-commonjs": "21.0.1",
     "@rollup/plugin-json": "4.1.0",
     "@rollup/plugin-node-resolve": "13.0.6",
     "@rollup/plugin-replace": "3.0.0",
-    "@vue/test-utils": "1.2.2",
-    "babel-core": "6.26.3",
+    "@vue/compiler-sfc": "^3.2.21",
+    "@vue/test-utils": "2.0.0-rc.16",
+    "@vue/vue3-jest": "27.0.0-alpha.3",
+    "axios": "0.24.0",
+    "babel-core": "7.0.0-bridge.0",
     "babel-jest": "27.3.1",
-    "babel-preset-env": "1.7.0",
+    "babel-preset-env": "^1.7.0",
     "bootstrap": "3.4.1",
     "cross-env": "7.0.3",
     "es6-object-assign": "1.1.0",
@@ -62,25 +64,24 @@
     "eslint-plugin-vue": "8.0.3",
     "jest": "27.3.1",
     "jest-serializer-vue": "2.0.2",
+    "less": "4.1.2",
+    "less-loader": "7.3.0",
     "lodash": "4.17.21",
+    "patch-vue-directive-ssr": "^0.0.1",
     "prettier": "2.4.1",
     "rollup": "2.60.0",
     "rollup-plugin-filesize": "9.1.1",
-    "rollup-plugin-strip-code": "0.2.7",
+    "rollup-plugin-strip-code": "^0.2.7",
     "rollup-plugin-uglify": "6.0.4",
-    "rollup-plugin-vue": "5.1.9",
+    "rollup-plugin-vue": "6.0.0",
     "sinon": "12.0.1",
-    "vue": "2.6.14",
-    "vue-jest": "3.0.7",
-    "vue-template-compiler": "2.6.14",
-    "@vuepress/plugin-active-header-links": "1.8.2",
-    "@vuepress/plugin-google-analytics": "1.8.2",
-    "less": "4.1.2",
-    "less-loader": "7.3.0",
-    "vuepress": "1.8.2",
-    "vuepress-plugin-redirect": "1.2.5",
-    "vuepress-plugin-sitemap": "2.3.1",
-    "axios": "0.24.0"
+    "vitepress": "^0.20.1",
+    "vue": "^3",
+    "vue-i18n": "^9.2.0-beta.17",
+    "vue-jest": "5.0.0-alpha.10"
+  },
+  "peerDependencies": {
+    "vue": "^3"
   },
   "jest": {
     "setupFilesAfterEnv": [
@@ -101,7 +102,7 @@
       "vue"
     ],
     "transform": {
-      ".*\\.(vue)$": "vue-jest",
+      ".*\\.(vue)$": "@vue/vue3-jest",
       "^.+\\.js$": "<rootDir>/node_modules/babel-jest"
     },
     "snapshotSerializers": [
@@ -128,7 +129,7 @@
     "Directives",
     "Vue",
     "VueJS",
-    "Vue2",
+    "Vue3",
     "WebComponents",
     "vue-strap",
     "vuestrap",
