@@ -1,7 +1,7 @@
 export { default as Carousel } from './carousel/Carousel.vue'
 export { default as Slide } from './carousel/Slide.vue'
 export { default as Collapse } from './collapse/Collapse.js'
-export { default as Dropdown } from './dropdown/Dropdown.js'
+export { default as Dropdown } from './dropdown/Dropdown.vue'
 export { default as Modal } from './modal/Modal.vue'
 export { default as Tab } from './tabs/Tab.vue'
 export { default as Tabs } from './tabs/Tabs.vue'
@@ -9,19 +9,19 @@ export { default as DatePicker } from './datepicker/DatePicker.vue'
 export { default as Affix } from './affix/Affix.vue'
 export { default as Alert } from './alert/Alert.vue'
 export { default as Pagination } from './pagination/Pagination.vue'
-export { default as Tooltip } from './tooltip/Tooltip.js'
-export { default as Popover } from './popover/Popover.js'
+export { default as Tooltip } from './tooltip/Tooltip.vue'
+export { default as Popover } from './popover/Popover.vue'
 export { default as TimePicker } from './timepicker/TimePicker.vue'
 export { default as Typeahead } from './typeahead/Typeahead.vue'
-export { default as ProgressBar } from './progressbar/ProgressBar'
-export { default as ProgressBarStack } from './progressbar/ProgressBarStack'
-export { default as Breadcrumbs } from './breadcrumbs/Breadcrumbs'
-export { default as BreadcrumbItem } from './breadcrumbs/BreadcrumbItem'
-export { default as Btn } from './button/Btn'
-export { default as BtnGroup } from './button/BtnGroup'
-export { default as BtnToolbar } from './button/BtnToolbar'
+export { default as ProgressBar } from './progressbar/ProgressBar.vue'
+export { default as ProgressBarStack } from './progressbar/ProgressBarStack.vue'
+export { default as Breadcrumbs } from './breadcrumbs/Breadcrumbs.vue'
+export { default as BreadcrumbItem } from './breadcrumbs/BreadcrumbItem.vue'
+export { default as Btn } from './button/Btn.vue'
+export { default as BtnGroup } from './button/BtnGroup.vue'
+export { default as BtnToolbar } from './button/BtnToolbar.vue'
 export { default as MultiSelect } from './select/MultiSelect.vue'
 export { default as Navbar } from './navbar/Navbar.vue'
-export { default as NavbarNav } from './navbar/NavbarNav'
-export { default as NavbarForm } from './navbar/NavbarForm'
-export { default as NavbarText } from './navbar/NavbarText'
+export { default as NavbarNav } from './navbar/NavbarNav.vue'
+export { default as NavbarForm } from './navbar/NavbarForm.vue'
+export { default as NavbarText } from './navbar/NavbarText.vue'
