@@ -1,5 +1,5 @@
-import Popover from '../../components/popover/Popover.js'
-import Vue from 'vue'
+import { h, render } from 'vue'
+import Popover from '../../components/popover/Popover.vue'
 import { hasOwnProperty } from '../../utils/object.utils'
 
 const INSTANCE = '_uiv_popover_instance'
@@ -7,55 +7,58 @@ const INSTANCE = '_uiv_popover_instance'
 const bind = (el, binding) => {
   // console.log('bind')
   unbind(el)
-  const Constructor = Vue.extend(Popover)
-  const vm = new Constructor({
-    propsData: {
-      target: el,
-      appendTo: binding.arg && '#' + binding.arg,
-      title:
-        binding.value && binding.value.title && binding.value.title.toString(),
-      positionBy:
-        binding.value &&
-        binding.value.positionBy &&
-        binding.value.positionBy.toString(),
-      content:
-        binding.value &&
-        binding.value.content &&
-        binding.value.content.toString(),
-      viewport:
-        binding.value &&
-        binding.value.viewport &&
-        binding.value.viewport.toString(),
-      customClass:
-        binding.value &&
-        binding.value.customClass &&
-        binding.value.customClass.toString(),
-    },
-  })
   const options = []
   for (const key in binding.modifiers) {
     if (hasOwnProperty(binding.modifiers, key) && binding.modifiers[key]) {
       options.push(key)
     }
   }
+  let placement, trigger, enterable
   options.forEach((option) => {
     if (/(top)|(left)|(right)|(bottom)/.test(option)) {
-      vm.placement = option
+      placement = option
     } else if (/(hover)|(focus)|(click)/.test(option)) {
-      vm.trigger = option
+      trigger = option
     } else if (/unenterable/.test(option)) {
-      vm.enterable = false
+      enterable = false
     }
   })
-  vm.$mount()
-  el[INSTANCE] = vm
+  const vNode = h(Popover, {
+    target: el,
+    appendTo: binding.arg && '#' + binding.arg,
+    title:
+      binding.value && binding.value.title && binding.value.title.toString(),
+    positionBy:
+      binding.value &&
+      binding.value.positionBy &&
+      binding.value.positionBy.toString(),
+    content:
+      binding.value &&
+      binding.value.content &&
+      binding.value.content.toString(),
+    viewport:
+      binding.value &&
+      binding.value.viewport &&
+      binding.value.viewport.toString(),
+    customClass:
+      binding.value &&
+      binding.value.customClass &&
+      binding.value.customClass.toString(),
+    enterable,
+    placement,
+    trigger,
+  })
+
+  const container = document.createElement('div')
+  render(vNode, container)
+  el[INSTANCE] = container
 }
 
 const unbind = (el) => {
   // console.log('unbind')
-  const vm = el[INSTANCE]
-  if (vm) {
-    vm.$destroy()
+  const instance = el[INSTANCE]
+  if (instance) {
+    render(null, instance)
   }
   delete el[INSTANCE]
 }
@@ -67,4 +70,4 @@ const update = (el, binding) => {
   }
 }
 
-export default { bind, unbind, update }
+export default { mounted: bind, unmounted: unbind, updated: update }
