@@ -3,19 +3,19 @@ import * as directives from './directives'
 import * as services from './services'
 import locale from './locale'
 
-const install = (Vue, options = {}) => {
+const install = (app, options = {}) => {
   // Setup language, en-US for default
   locale.use(options.locale)
   locale.i18n(options.i18n)
   // Register components
   Object.keys(components).forEach((key) => {
     const _key = options.prefix ? options.prefix + key : key
-    Vue.component(_key, components[key])
+    app.component(_key, components[key])
   })
   // Register directives
   Object.keys(directives).forEach((key) => {
     const _key = options.prefix ? options.prefix + '-' + key : key
-    Vue.directive(_key, directives[key])
+    app.directive(_key, directives[key])
   })
   // Register services
   Object.keys(services).forEach((key) => {
@@ -24,7 +24,7 @@ const install = (Vue, options = {}) => {
       const _key = options.prefix
         ? options.prefix + '_' + serviceKey
         : serviceKey
-      Vue.prototype['$' + _key] = service[serviceKey]
+      app.config.globalProperties['$' + _key] = service[serviceKey]
     })
   })
 }
