@@ -1,27 +1,19 @@
-import * as uiv from './install'
+import * as uiv from './index'
+import { createApp } from 'vue'
 
 describe('install', () => {
-  function Vue() {
-    // nothing
-  }
-
-  Vue.component = (key) => {
-    // console.log(key)
-  }
-  Vue.directive = (key) => {
-    // console.log(key)
-  }
-
-  let cSpy, dSpy
+  let cSpy, dSpy, app
 
   beforeEach(() => {
-    cSpy = jest.spyOn(Vue, 'component')
-    dSpy = jest.spyOn(Vue, 'directive')
+    app = createApp({})
+
+    cSpy = jest.spyOn(app, 'component')
+    dSpy = jest.spyOn(app, 'directive')
   })
 
   it('should be able to install with prefix', () => {
     // simulate a Vue.use
-    uiv.install(Vue, { prefix: 'uiv' })
+    app.use(uiv, { prefix: 'uiv' })
     // components
     expect(cSpy).toBeCalledWith('uivAlert', expect.any(Object))
     expect(cSpy).toBeCalledWith('uivModal', expect.any(Object))
@@ -29,7 +21,9 @@ describe('install', () => {
     expect(dSpy).toBeCalledWith('uiv-tooltip', expect.any(Object))
     expect(dSpy).toBeCalledWith('uiv-scrollspy', expect.any(Object))
     // methods
-    expect(Vue.prototype.$uiv_alert).toEqual(expect.any(Function))
-    expect(Vue.prototype.$uiv_notify).toEqual(expect.any(Function))
+    expect(app.config.globalProperties.$uiv_alert).toEqual(expect.any(Function))
+    expect(app.config.globalProperties.$uiv_notify).toEqual(
+      expect.any(Function)
+    )
   })
 })
