@@ -1,26 +1,19 @@
 import { TYPES } from '../../constants/messagebox.constants'
-import { removeFromDom } from '../../utils/dom.utils'
-import { spliceIfExist } from '../../utils/array.utils'
 import {
   isFunction,
   isExist,
   isString,
   isPromiseSupported,
-  assign,
 } from '../../utils/object.utils'
 import MessageBox from '../../components/messagebox/MessageBox.vue'
-import Vue from 'vue'
+import { render, h } from 'vue'
 
-const queue = []
-
-const destroy = (instance) => {
+const destroy = (container) => {
   // console.log('destroyModal')
-  removeFromDom(instance.$el)
-  instance.$destroy()
-  spliceIfExist(queue, instance)
+  render(null, container)
 }
 
-// handel cancel or ok for confirm & prompt
+// handle cancel or ok for confirm & prompt
 const shallResolve = (type, msg) => {
   if (type === TYPES.CONFIRM) {
     // is confirm
@@ -32,37 +25,34 @@ const shallResolve = (type, msg) => {
 }
 
 const init = function (type, options, cb, resolve = null, reject = null) {
-  const i18n = this.$i18n
-  const instance = new Vue({
-    extends: MessageBox,
-    i18n,
-    propsData: assign({}, { type }, options, {
-      cb(msg) {
-        destroy(instance)
-        if (isFunction(cb)) {
-          if (type === TYPES.CONFIRM) {
-            shallResolve(type, msg) ? cb(null, msg) : cb(msg)
-          } else if (type === TYPES.PROMPT) {
-            shallResolve(type, msg) ? cb(null, msg.value) : cb(msg)
-          } else {
-            cb(msg)
-          }
-        } else if (resolve && reject) {
-          if (type === TYPES.CONFIRM) {
-            shallResolve(type, msg) ? resolve(msg) : reject(msg)
-          } else if (type === TYPES.PROMPT) {
-            shallResolve(type, msg) ? resolve(msg.value) : reject(msg)
-          } else {
-            resolve(msg)
-          }
+  // const i18n = this.$i18n
+  const container = document.createElement('div')
+  const vNode = h(MessageBox, {
+    type,
+    ...options,
+    cb(msg) {
+      destroy(container)
+      if (isFunction(cb)) {
+        if (type === TYPES.CONFIRM) {
+          shallResolve(type, msg) ? cb(null, msg) : cb(msg)
+        } else if (type === TYPES.PROMPT) {
+          shallResolve(type, msg) ? cb(null, msg.value) : cb(msg)
+        } else {
+          cb(msg)
         }
-      },
-    }),
+      } else if (resolve && reject) {
+        if (type === TYPES.CONFIRM) {
+          shallResolve(type, msg) ? resolve(msg) : reject(msg)
+        } else if (type === TYPES.PROMPT) {
+          shallResolve(type, msg) ? resolve(msg.value) : reject(msg)
+        } else {
+          resolve(msg)
+        }
+      }
+    },
   })
-  instance.$mount()
-  document.body.appendChild(instance.$el)
-  instance.show = true
-  queue.push(instance)
+  render(vNode, container)
+  document.body.appendChild(container.firstElementChild)
 }
 
 // eslint-disable-next-line default-param-last
